Refactored to clean up some duplicated code and wrote a few more populate db methods. I also fixed a few bugs and added more logging.

+++ DB Schema/NFL_DB_Schema.sql
@@ -52,7 +52,7 @@ CREATE TABLE PlayerStats (
 	playerId INT UNSIGNED,
 	statId INT UNSIGNED,
 	seasonID INT UNSIGNED,
-	statValue INT NOT NULL,
+	statValue DECIMAL NOT NULL,
 	FOREIGN KEY (playerId) 
 			REFERENCES Players(playerId)
 			ON UPDATE CASCADE ON DELETE CASCADE,
